fix ethernet to integer conversion
we can't really memcpy() an ethernet address to an integer64
attribute, because the address may be in the high bits of the
integer, or in the lowe bits, depending on endian-ness of the
host CPU.

Instead, we manually copy the bytes over, so that we're sure they
end up where we want.  Note that this is the same process used
by fr_value_box_cast_to_ethernet(). i.e. copy to the *low* bytes,
and not to the *high* bytes.

In addition, fix the unit tests to match.  Casting a string type
to ethernet isn't the right thing to do.  With the previous code,
the string type contained an ASCII representation of the ethernet
copied to the high bytes of integer64 and then printed... which
isn't at all anything useful.

The new test case prints ethernet to integer64 type, and then
casts that to ethernet.

If we want to cast from ethernet to string and back, we MUST
just copy the bytes raw / as-is, and *not* convert them to integer