return attr_len, not rcode
i.e. even if the called function decodes *less* than the attribute,
we tell the caller that it decoded *all* of the attribute.

This change ensures that if a function partially decodes an attribute,
that the caller will skip *all* of the attribute.  And not start
decoding the next attribute in the middle of the partially decoded
one.