disallow struct[n]
which means that we don't really care about the size of the
structs.

On encoding, if we run out of room, we just truncat the data
as with every other data type.

On decoding, if there's too little data, we just return what
data we can find.  If there's too much data, we just ignore
the extra bits.