convert detail reader to thread-safe.
And fix bug where when mod_write() unpauses the socket, it also
seeks to the start of the file.  This ensures that if there is
leftover data in the buffer, mod_read() will be called again.

We can't rely on network.c to check "leftover", because it has
no idea if the socket "no data ready" means we should still call
mod_read().  For detail, yes.  For other protocols, maybe no.

Also fix the work instantiation so that it's done once, and not
multiple times.