Show when log config has errors and don't quit on deprecated items
As parsing the log section happens before logging is configured,
any errors that happen don't get displayed. If logging has been
configured on the command line then the log section isn't read at
all, so again the errors are not found. Set the log destination
temporarily so that the errors are printed to stderr.

cf_section_parse returns -2 for deprecated items, so behave in the
same way as the rest of the config and just ignore them.