set dl_loader = NULL only if the trees are empty
and add assertions that dl_loader!=NULL elsewhere

It appears that in 'tests.eap', something is freeing the dl_loader
after the server starts, but before the server has exited.  The result
is that it would crash instead of doing proper cleanups.

This change at least makes it keep running.  But the underlying
issue needs to be tracked down.