Update systemd units to depend on network-online
Due to the use of SO_BINDTODEVICE, FreeRADIUS requires the address to be
up when binding to a specific IP address. In systemd, depending on
network.target is not sufficient to require the network to be up, so
FreeRADIUS can still fail to bind; instead, network-online.target is
required if the network is required to be online.

This manifests in errors such as:

Error: Failed binding to auth address x.x.x.x port 1812 bound to
       server default: Cannot assign requested address

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1637275

Signed-off-by: Alexander Scheel <ascheel@redhat.com>