mark up autoloaded dictionaries vs manual ones
So that dictionaries can be manually loaded from raddb/dictionary,
and we can later still autoload the various protocol dictionaries

The only downside is that src/tests/Makefile still fails.
The server loads the test dictionary and parses the config files
correctly.  radclient loads the test dictionary and sends the
attributes correctly.

But somehow (magically) when the server receives the packets, it
can't find the test vendor, and prints out all of the attributes
as raw ones...