don't artificially truncate debug output.  Fixes #2381
The buffer size for VPs being printed should be as long as
the attribute name, plus spaces, plus operator, plus quotes,
plus encoded content.  In the worst case, where a printable
string contains all binary data, it will be 4x the length of
the string.  So we'll need ~64 + 6 + 4*247, or a bit more than
1K of room.

That's pathological, so we don't care so much about it.
Instead, increase the size of the buffer so that common cases
aren't truncated.

Note also that we print out the *full* string, even if the
encoded version is truncated to 257 octets.  There's currently
no good way to expose that, unfortunately.  We would need to
add a callback to the *encoding* routine which took a print
function, and could then pass the *encoded* length to the print
function.