apply options from the bottom up
i.e. if an option is at depth N, and the same option is at depth N+1,
then we apply the option at depth N+1, and ignore the option at
depth N.

We do this the naive way.  By adding options from the bottom up,
but ONLY if the options don't already exist in the reply.
i.e. adding options *after* we've recursed into the child
sections.

This ends up being maybe O(N^2) in the number of options.  But
the number of options is small, so oh well...