Fix TrustRouter query potential deadlock
Existing code uses a single lock to protect concurrent access to the
global_tidc variable. Under certain circumstances, two FreeRadius servers
wanting to resolve each other's realms at the same time might provoke
a deadlock situation, if this query triggered in turn another realm
resolution attempt for the APC's realm on both sides.

While this is a corner case, it has been observed in stress testing
environments.

This patch gets rid of the global_tidc in favour of a local approach,
and provides a pool of mutexes that are used to serialise different queries for
a particular realm, while allowing concurrent queries for different realms.