fix untagged tunnel encrypted attributes in CoA requests
When sending CoA, disconnect and accounting requests we unconditionally
added a tag to any tunnel encrypted attributes regardless of the "has_tag"
flag. The length was not adjusted, which made the error hard to distinguish
from a password mismatch.

Fix by refactoring, avoiding unnecessary code path differences for the
different packet types.

Fixes: c5aa669eeb66 ("Simplify RFC format attributes")
Signed-off-by: Bjørn Mork <bjorn@mork.no>