add "accept_conflicting_packets" config
as a first stage of negotiating request authenticator as an
extended ID.

For now, we can accept these conflicting packets, and likely add
a `do_not_respond` policy.

We will later update the code to include an attribute that holds
the Original-Request-Authenticator, and add that to the reply.

And then add the ability in rlm_radius to negotiate it