support for calling any named method, not just "authorize", etc.
the module_by_name_and_method() now returns a component *and*
name1 / name2 of the named method.  e.g. `sql.authorize` or
even `sql.session.start`.

It also returns the method function to call.  This is so that
the compiler doesn't need to troll through the module method list.
Instead, it's all buried inside of one lookup function.