don't necessarily "" quote strings with '.' in them.
This is to work around the new requirement that dictionary
attribute names can't have '.' in them.  But, the array of allowed
characters is used by the tmpl code to determine if it quotes
a string, so we've got to hack this up to counter the dictionary
changes.

Since the tmpl hack is already marked as a hack, it's better to
have hacks all in one place