bootstrap modules AFTER virtual server bootstrap
possibly a temporary change.  There are a number of modules
which *define* aliases for Auth-Type.  This is entirely wrong.
The names for the Auth-Type aliases are taken from the virtual
servers, in the 'authenticate foo' sections.  The modules should
*look* for the named alias, and complain if it doesn't exist.

Simply defining the aliases in the modules means that we can have
run-time errors due to a bad config.

Plus, alias lookups are case insensitive, while CONF_SECTION name
lookups are case sensitive.  So there are more chances for errors.