add MEMBER keyword to dictionaries for 'struct'
Update the "man" page, update the formatter, update the default
dictionaries.

For now, the parent of MEMBER is implicit in the order of the
dictionary file.  i.e. we don't have a reference as with VALUEs

since there isn't any need to place MEMBERs randomly through the
file, this limitation is acceptable.