just box a normal string, not a tainted one.
the previous commit fixes the printing routine to always escape
strings when they're being printed.

When tokenizing xlat strings, we presume (for now) that the input
is always safe.  i.e. we don't allow %{%{User-Name}}.

If we do need to allow xlat parsing from tainted sources, then the
entire xlat API needs to be updated to mark the input with a
tainted flag