clean up and finalize use of ctx->stack
so that "previous_attr" becomes "relative_attr".

We also add a dict_ctx_push() function which pushes attributes
onto ctx->stack.

And dict_ctx_unwind() for unwinding the stack multiple frames
when necessary.

There are likely edge cases which aren't handled well.  The code
sometimes does an *implicit* push of struct/tlv.  And other times
an *explicit* push / pop.  The differences are not tracked, which
makes it more difficult to automatically determine what to do.