add break_point and return_point to interpreter
top frames are now both, just for sanity.

break points are "foreach" instructions.  Which SHOULD be
verified by the compiler, but it doesn't hurt to have a sanity
check.

return points are (for now) only top frames.  But marking them
up this way means that we may be able to have return points which
are in policies