print warnings only when asked.  Fixes #3032
It will now ignore warnings in normal builds, and only them
when running:

    make check.configure

this change is because we can't control the timestamps on the
`configure` and `configure.ac` files on `git clone` or
`git checkout`.  The `git` program may write one file before
another, and not necessarily in any order we would like.

Since we can't fix that, we just suppress the warnings.