The 'Domain-List' value should be string
As described in https://tools.ietf.org/html/rfc1035#section-3.1

Domain names in messages are expressed in terms of a sequence of labels.
Each label is represented as a one octet length field followed by that
number of octets.  Since every domain name ends with the null label of
the root, a domain name is terminated by a length byte of zero.

i.e: It should be encoded as a 'dns_label' to append the '\0'