set default type size / length in fr_dict_t
so that the dict tokenizer uses that instead of hard-coded '1'.

And change the BEGIN-VENDOR to use "parent=..." instead of
"format=..".  As the reference is to a parent attribute, NOT
to a particular format.

The VSA encoding flags should really be part of the BEGIN-VENDOR
line, as that is protocol-specific.  Right now they're in the
VENDOR, which unfortunatelty is used across multiple protocols.

We probably want to update the code to look for encoding sizes
in the VENDOR attribute, and not in the fr_dict_vendor_t