start of autoload and validation for dictionaries.
if we load protocol FOO, we now look for libfreeradius-FOO
which *should* already be loaded.  i.e. when the application links
to libfreeradius-radius and calls fr_radius_init(), that function
loads the RADIUS dictionaries.  Which now link back to the library.

The dictionary code then looks for a symbol called

	libfreeradius_FOO_dict_protocol

and saves a pointer to that structure in the dictionary.  We can
now have protocol-specific validation functions.