Add many more tests
Add cancel partial state for dealing with partial writes when cancelling

Restore in_handler correctly when doing nested handler calls.  We *may* need to move this to a stack, but lets not do that unless we have to.

When requing requests, free the cancelled ones, don't try and move them to a new connection *sigh*