Change how we deal with freeing connections
We now use the on_halt watcher as an indication the conn is ready to be freed, and that the tconn should also be freed.

This allows us to signal that the connection should be shutdown, and allows that process to complete before we attempt to free the conn or tconn.

The previous code just failed spectacularly... this is almost sane.