add local free list for MD5 contexts
which saves 1/2s to 1s on the performance tests.

Because we don't control the content of OpenSSL's structures,
we can't use a dlist.  Instead, we just have a fixed-size local
array.  With provisions for all array entries being used.

In general, the code should allocate an MD5 context, use it,
and immediately free it.  i.e. *without* doing any IO while
holding the MD5 context.

Since the code does this already, it's OK.