Minor cleanups in channel code
Use REQUESTOR in place of "master" and "network", and RESPONDER in place of "worker". The channel code will likely be reused very soon to create module specific thread pools to deal with blocking requests, and it's confusing to use usecase specific terms in generic code.

Move all size_t to uint64_t.  Guessing the original author was just being lazy and didn't want to type %"PRIu64 in a bunch of places.

Rename "ctx" to "uctx" because that's what it is...