add "to do" list for timers, too
New timers were added while running timers. Instead of
being added to the main heap, they were added to the
"in progress" list.  We can now add them to the main
heap.

Doing it this way prevents the server from running
into an infinite loop when it takes a long time to
service FD events.  The timer callback MAY add a new
timer which is in the past.  So the above loop could
run forever.