fix use of mod_rcode_table
the definition for it shouldn't be in src/lib/server, but in
src/lib/unlang, where it's defined.

Since that definition has move, update the rest of the server
to use the src/lib/server "rcode_table", instead of the
unlang-specific "mod_rcode_table".

We should probably just get rid of the mod_rcode_table entirely,
unless we want unlang tests which don't include any server
functionality