move sub-config to in-line struct
the configuration parser for sub-structures / sub-sections
does NOT take a pointer to a pointer.  Instead, it takes a pointer
to a struct.  So we need to include the struct in-line.

Plus, the config parser is called after the module instance is
allocated, but before the module can do anything to it.  So the
trunk code cannot contain a pointer to a sub-struct, because
there is no way to properly initialize the pointer.