Remove treq and request from the udp_request
Fix parenting of the udp_request.  Should be parented from the treq because its lifetime is bound to the treq, not to the REQUEST *.

It doesn't _really_ matter here, but it does mean we can take advantage of the talloc_pool/freelist in the trunk.

Where it does matter is for trunk protocols which do cancellations, as the REQUEST * and the rctx might go away, but the preq needs to live on so we can cancel any upstream queries.