use real upsert (ON CONFLICT) instead of workaround in postgresql cui
Clause ON CONFLICT performs better than the RULE workaround and it's
guaranted to be atomic (the workaround is not).

It requires PostgreSQL 9.5 or newer; this version was released in 2016
and the previous versions are already unsupported, so it shouldn't be
a problem for anyone using latest FreeRADIUS versions.

There's a problem with backward compatibility because ON CONFLICT
doesn't work when there's a RULE ON INSERT defined on the table. It
results in the following error in radius.log:

    Error: rlm_sql_postgresql: 0A000: FEATURE NOT SUPPORTED