add API / wrapper fro fr_pair_group_...() functions
so that we can do bulk conversions of the callers *without*
breaking anything.

i.e. convert the callers to use the new API, which (for now)
is just a macro wrapper around the old API.  Then, once the callers
use the new API format, we can delete the macros, and actually use
the real new API functionality.