use random padding, instead of zeros
that way if the salts are the same, the encoded data will still
be all different.

RFC 2868 Section 3.5 recommends that the padding is all zeros.
But the field already contains a length, so it's not necessary.

And for CoA-Request packets, the "encyption" of the password
is done via the request authenticator, which is all zeros.
So we rely only on the salt for security.  Which is small, and
therefore terrible.

Changing to random padding makes it harder to figure out what's
going on.