rlm_icmp: Fix icmp6 support
The documentation says that sockets() created with IPPROTO_ICMPV6 will receive packets
without the IPv6 header nor IPv6 extension headers.

In this case, we don't have a way to extract the source address. That is the reason for
we are using the ICMP requests 'ident' field instead of the IP address.