move GNU make magic && shell scripts to more GNU make magic
the rules to create $(OUTPUT)/%.attrs didn't work.  This failure
is shown by the fact that the make file has 3 different ways
to produce the same files:

* a set of BOOTSTRAP variables
* $(OUTPUT)/%.attrs: %(DIR)/%.attrs rule
* shell script if / cp rules

So we delete all 3 methods, and replace them with a simple 'cp',
along with a $(if) and $(wildcard), to select files A or B.

We would _like_ to do:

$(OUTPUT)/%: $(DIR)/% $(OUTPUT)/%.attrs

except GNU make it too stupid to allow that.