add function to turn unlang ok / fail to reply packet code
in reality, this function just gets asked ok / fail as a bool, and
returns the appropriate packet code.

This is for use only by unit_test_module.  That program runs
unlang, and then needs to set the correct reply packet type.
This reply packet type depends both on the request packet type,
and on the unlang status ok / fail.

In order to make unit_test_module protocol agnostic, we export
a function in the RADIUS library to help it out.