switch to using virtual_server_get_process_by_name()
instead of unlang_interpret_synchronous()

This change means that we have to run our own event loop.  But,
it also means that we just call the proto modules to do all of
the processing.  Which means that unit_test_module no longer
needs to know about any magic packet codes.  The proto_radius_auth
etc. functions take care of all of that.