New helper scripts for defining IP pools and aligning sqlippools
generate_pool_addresses.pl allow multiple pools, each containing one or more
address ranges (IPv4 or IPv6) to the specified by YAML file and will output a
list of addresses meeting the specification.

These address ranges may be defined as contiguous or sparse (useful in the case
of IPv6!). In the case of sparse ranges, if a file listing pre-existing
addresses is supplied then these are used. The remainder of the addresses (up
to the given capacity of the range) are allocated randomly.

align_sql_pools.pl takes the list of addresses output by
generate_pool_addresses.pl and generates a set of dialect-specific SQL
statements to align the sqlippools, preserving any existing leases that are
using an addresses that is retained in the pool.