clean ups and fixes.
hoist the "wait for PID" call to immediately after the fork.
The previous code waited until the process was done writing,
which meant that the waitpid was sometimes missed.

Add "state->failed" so that we can distinguish administrative
failure (timeout or too much output) from program failure.

allow the kill timer to run until waitpid() returns. This change
lets the program continue to run after closing it's STDOUT.
But also enforces that the program does not run forever.

distinguish "no more data" from "too much data".  When there's
no more data, we close the pipe and remove the IO handler.
But we let the program keep running.

When there's too much data, we remove the reader callback, close
the pipe FD, kill the program, and set the "failed" status.