implement dedup.
the old master.c never called rbtree_insert().  Oops.

The new code does that, and also abstracts the dedup parameters.
There is now a new function which creates a dedup tracking structure.
For RADIUS this is the first 20 bytes of the packet.  For DHCPv4
it is the XID, CHADDR, and message type code.  Having a separate
structure means that we don't have to search through the entire
packet every time we do rbtree comparisons.