Handle SQLITE_BUSY as an error
Exclusive locks can cause SQLite to return an SQLITE_BUSY error.
Treating this as an error causes processing of the current packet to be
abandoned complete with clearing up of the query.  Previously it was
being treated as a need for a reconnect which was causing a memory leak
as the existing connection was not cleared before reconnecting.