add tacacs_length()
which returns the minimum length necessary for a packet to be
passed to fr_tacacs_decode().

Note that the packet may still be invalid, because TACACS+ is a
crap protocol with almost random fields in random packets.  We
don't want to duplicate decode / validation logic in multiple places.

So even if fr_tacacs_length() returns that the buffer/buffer_len
is a TACACS+ packet, we may still have fr_tacacs_decode() return
"invalid packet".