create and use fr_network_config_t
for max_outstanding, which mirrors max_queue_size in v3.

For now, if we try to send a packet to a worker and it's too busy,
we just drop the packet.  i.e. we don't try to send it to another
worker.

The configuration option is also not set, as the config parsing
isn't yet plumbed in.