dhcp_sqlippools: Don't handle a DHCP REQUEST like a DISCOVER
DHCP REQUESTs are now ACKed or NAKed based on whether or not the address
requested by the client was actually offered during a previous DISCOVER.

Previously the IP allocate_find process was invoked for both DISCOVER and
REQUEST which was unnecessary and could result in an ACK of an address which
wasn't actually requested by the client in some circumstances (e.g. when there
are multiple active offers made by racing servers).

Additionally, any other actively offered addresses made to the client
(by other servers sharing the same database) that are not currently being
requested are immediately returned to the pool.