sqlippools: Support for requested IP address hint
Add req_attribute_name configuration item that specifies which request
attribute is to be used as a hint for the IP address to assign.

For RADIUS this is likely to be the Framed-IP-Address (IPv4) or
Framed-IPv6-Prefix (IPv6) attributes sent during Access-Request processing.

For DHCP this will be the DHCP-Requested-IP-Address parameter, usually sent in
renewals and during rebinding.

It is only a hint. It is passed to a new allocate_requested query (if defined)
which will allocate the address if it is free, AFTER first running the
allocate_existing query (if defined) that allocates the device's existing or
previous address. This prevents a rogue device using up multiple pool entries,
and is per RFC 2131 for DHCP:

  * Issue current address
  * Else, issue previous (now expired) address, if free
  * Else, issue requested address, if free
  * Else, issue a new address from the pool of free addresses

Query updates for the various databases to follow...