Add alloc_requested and update_free queries
alloc_requested is an optional query that can be used to check for
availability of a requested IP address (as can happen in DHCP Discover).

update_free is an optional query that can be used to tidy up prior to
updating leases.  Primarily intended for multi-server DHCP setups
sharing a common SQL back end which could result in multiple pool
addresses being offered to a client.