Add counter to DHCP ippool schema
By updating a counter every time a REQUEST is responded to successfully
we ensure that records are changed, thus all database engines return an
affected row count of 1 for successful matches. This safeguards the new
DHCP site config which uses the affected row count to determine between
ACK and NAK responses.