Make marker pointer advances past parent pointer advance parent pointer too
Why? fr_[sd]buff_moves() with marker source/destination will advance the
marker pointer... but if that moves it past the parent's pointer, then
bytes are used that aren't reflected in fr_[sd]buff_used(). For dbuffs,
at least, this breaks the convention for encoding functions.

This change causes sets of marker pointers to check the parent's pointer
and, if and only if the set moves the marker pointer past the parent
pointer, advance the parent pointer to match the marker pointer.