allow subrequest { ... }
and remove code which allowed it to specify "component".
None of the tests checked for that.  The documentation didn't
mention it.  And, I have no idea what it would mean to call
"recv Access-Request", but with component "accounting".

So the code was removed.

Now if there is no dictionary / packet name, it will use the
current dictionary, and request->packet->code in order to
initialize the child.