Use trie for "closest enclosing subnet" matching of addresses in rlm_files
A UNION fr_rlm_files_tree_t is used to represent the lookup tree as either
an RB tree or a patricia trie.

If the data type of the key is an IPv4/6 address or prefix then use
the trie entry to store the lookup data.  Lookups into this will then return
the closest enclosing subnet match.