decode tagged attributes
through the (cough) simple expedient of defining a set of magic
attributes "Tag-1" through "Tag-31", of data type "group".

This will horrify anyone who is used to 20+ years of RADIUS
configuration.  But there's little else we can do in v4, without
butchering a lot of the code.

So, the poor admins will have the cost of changing things.  Sorry.