fix up dhcpv4 decoder.
Options are allowed to have zero length.  In which case they
MUST be "octets" or "string".  Or for other data types, they MUST
be converted to "raw".

Fix up the decode_value_internal() function to skip calling
fr_value_box_from_network() if there's no data.

Fix the callers to allow it to return 0 data.

fix fr_dhcpv4_decode_proto() to return the total number of bytes
decoded, instead of "0" for success.

Update and add more tests for error conditions