move ascend filters to pointer, and fix use of vb_length
vb_length _was_ inside of the union for variable-sized types,
but the code also set it for fixed-size types.  Only the size
of the ascend filter prevented various over-writes.

changing ascend filter to pointer instead of an array highlighted
this issue, and causes us to move the "length" field out of the
union in order to stop the over-write