add FR_TYPE_NON_BLOCKING
for use with tmpl_eval().  We can mark configuration items as
allowing tmpls / xlats, BUT also mark them up with a requirement
for non-blocking operation.

This lets us convert much of the code to using tmpl_eval() instead
of xlat_eval().  At the same time, the admin is prevented from
breaking the server by using asychronous operations in places
which do not support them.

Once all of the callers of foo_eval() have been converted to
async, we can remove the FR_TYPE_NON_BLOCKING flag.