rlm_sql: Use the non-xlat escape function for the group membership query
Use the basic sql_escape_func when expanding the group membership query that
uses our existing SQL handle.

If we use sql_escape_for_xlat_func, which reserves a connection, then when the
connection pool is exhausted we erroneously conclude that the user is not a
member of any groups.