Add fr_dbuff_uint64v_out(), and float/double options for fr_dbuff_out()
Unlike fr_dbuff_uint64v_in(), one must pass a length between 1 and 8
inclusive to specify the number of bytes to read, as well as a dbuff
or marker and a pointer to a uint64_t. Bytes read are interpreted as
an unsigned integer in network order, converted to host order, and
stored in the pointed-at uint64_t.