a few fixes
check for valid tmpl types, and cast to correct types
so that the function calling map_afrom_sbuff() has to do less
work.

update map_assignment_table for attr_filter
we need assignments, but also filtering operators.

add parameter for terminating the RHS
which is only used when the RHS is a bare word.  Otherwise
we know when the RHS quoted string stops:  at the quote.