add encode / decode of "raw" attributes
something seems to not be set right.  We create

	raw.NAS-Port = 0x0102

but the NAS-Port is encoded as a *normal* NAS-Port
attribute, and encoded with 4 bytes instead of 2

set flags.length=0 for unknown attributes

because they're "octets", and fixed size.  So the encoder should
not use the length from the _old_ flags (e.g. 4 for uint32)

Instead, it should just use the length of the octets.

Also corrected the tests to match.