short-circuit TLV encoder for structs
if we're encoding a 'struct' and the last field is a TLV, then
the struct encoder can't encode it.  Because it needs to know
the sizes of "T" and "L", and it doesn't know that.

Also, if the 'struct' ends with a TLV, then the struct cannot
have a "key" field, as we don't (yet) encode sub-structs in the
middle of a struct.  Instead, the sub-struct has to be encoded
at the end of the parent struct.  Which means that the parent
can't end with a TLV.

The struct can't also have an auto-encoded 'length' field.
Because there's no way for it to be filled in with the length
of the base struct + TLVs.  If the struct_to_network() function
was passed an "encode_tlv" function, then this last restriction
could be relaxed.