create a private talloc ctx for tmpl_cursor_init()
because there are paths which create leaf nodes allocated from
this context, BUT those leaf nodes are never cleaned up.  Instead,
we rely on talloc() to clean them up when the parent talloc ctx
is cleaned up.

Except that most callers of tmpl_cursor_init() pass it NULL as the
talloc ctx.  Which means that this memory will NEVER be cleaned up.

In the interest of simple fixes, we fix the tmpl cursor code to
always create a child context, and always clean up that child
context.  Which is only 2 lines of code, and fixes all of this.