remove old functions and simplify
we now have cond_cast_tmpl(), which casts a tmpl_t, and resolves
it if necessary.  It also checks for IP addr -> prefix casting,
as that depends on the values, and can't be done in fr_type_promote()

We now also have cond_promote_types(), which checks LHS / RHS
for casts, ensures that they're compatible, and calls cond_cast_tmpl()
in order to cast the LHS / RHS tmpl to the appropriate type.

As a result, some unit tests have changed.  Generally more things
are now allowed, as the compiler is smarter.

The one (minor) regression is that we don't do compile-time checks
for things like (&Attr-uint8 > 256), which should arguably be a
parse error.  Instead, the RHS is a 16-bit value, so we promote
the LHS to a 16-bit value.  It ends up being a run-time check
which is always false.