resolve ALL tmpls before passing them to cond_eval()
which means that cond_eval() just has to "realize" the tmpls
before evaluating.  Which makes its life a lot easier.

We also add assertions to cond_eval() in order to force that
all of the callers have realized the tmpls.

It makes no sense to call cond_eval() on things which are
tmpl_is_unresolved().  *Someone* should have determined what that
thing is.