Reworked version of raidus_pairmove()
Rather than break the "to" and "from" lists into arrays, this
implementation keeps the fr_pair_list_t structures but uses arrays to
mark which "to" entries are edited / to be deleted.

New pair_list functions are used for list manipulation.

The original int counters are retained for debug messages.