clean up final uses of dict_attr_size
the struct validation code now checks if (a) the da contains
a known fixed size, or (b) it's a FR_TYPE_VALUE.  That covers
the same situations as before.

It still doesn't allow for variable-sized fields like dns labels,
but that will have to be another step.

now that dict_attr_size isn't used, it's removed.