Turn floating point comparison into memcmp() to silence compiler (#3849)
In numerical methods, it is indeed a bad idea to compare
floating point numbers for equality, but this just tests
that fr_dbuff_in() and fr_dbuff_out(), barring lack of space
breaking fr_dbuff_in(), are inverses for float and double.
No floating point arithmetic involved, so it's reasonable to
compare the raw bits. We'll use memcmp() to (1) emphasize that
and (2) keep the compiler from warning about comparing floating
point values for equality.