Get rid of the only explicit use of FR_DBUFF_RESERVE() (#3844)
The sole explicit use was in internal_encode(); it's not necessary
because in the case of the length requiring more than one byte,
fr_dbuff_move() will return error if there's not enough room. Also:

* fr_dbuff_move() can take advantage of extensible dbuffs
* by the time fr_dbuff_move() is called, we really know how much
  space is needed. Before, the code insisted on sizeof(uint64_t)
  bytes, corresponding to lengths unlikely in practice and thus
  possibly caausing a needless error return.

This change leaves FR_DBUFF_MAX() and FR_DBUFF_MAX_NO_ADVANCE()
expressed in terms of FR_DBUFF_RESERVE(), so FR_DBUFF_MAX() can
first be expressed directly and then move to a version that can
take advantage of extensibility.