Scheduled fuzzing improvements
Change scheduled fuzzing to a matrix workflow so that we fuzz protocols in separate
jobs: More CPUs; less context switching.

If HEAD doesn't build then walk up the tree until we find something that does,
then fuzz that.

Set the runtime based on the calculated time remaining in the job since we
don't know ahead of time how much time we will spend walking the tree to find a
working build.

Name assests after "<commit>:<protocol>"

Copy the job output text into the asset that is uploaded when there are fuzzing
failures.